App/requirements.txt
python==3.10
cryptography==39.0.1
pyperclip==1.8.2
pyserial==3.5
pillow
